recast output in uint8

+++ dlutils/preprocess.py
@@ -27,8 +27,7 @@ def unpreprocess(x, data_format,mode):
         im += 1.
         im *= 127.5
         im = np.clip(im, 0, 255)
-        im = im.astype(int)
-        return im
+        return im.astype(uint8)
 
     if mode == 'torch':
         mean = [0.485, 0.456, 0.406]
@@ -83,4 +82,4 @@ def unpreprocess(x, data_format,mode):
          
     im = np.clip(im, 0, 255)
 
-    return im.astype(int) 
+    return im.astype(uint8)
